requirements.txt
anytree>=2.12.1,<3.0.0
numpy>=1.26.2,<2.0.0
openai>=1.54.3,<2.0.0
pandas>=2.1.3,<3.0.0
pytz>=2024.2,<2025.0
regex>=2024.11.6,<2025.0
requests>=2.32.3,<3.0.0
sentence-transformers>=3.2.1,<4.0.0
tenacity>=9.0.0,<10.0.0
tiktoken>=0.7.0,<0.8.0
tqdm>=4.67.0,<5.0.0
vllm>=0.6.3.post1,<1.0.0
vertexai>=1.71.1,<2.0.0
anthropic>=0.39.0,<1.0.0
pyyaml
google
google-generativeai

pypdfium2 >= 4.30.1